+++ config/mercure.config.php
@@ -6,7 +6,7 @@ namespace Shlinkio\Shlink\Common;
 
 use Laminas\ServiceManager\AbstractFactory\ConfigAbstractFactory;
 use Lcobucci\JWT\Configuration;
-use Shlinkio\Shlink\Config\Factory\ValinorConfigFactory;
+use Psr\Container\ContainerInterface;
 use Symfony\Component\Mercure\Hub;
 
 return [
@@ -19,7 +19,15 @@ return [
             Mercure\LcobucciJwtProvider::class => ConfigAbstractFactory::class,
             Hub::class => Mercure\HubFactory::class,
             Mercure\MercureHubPublishingHelper::class => ConfigAbstractFactory::class,
-            Mercure\MercureOptions::class => [ValinorConfigFactory::class, 'config.mercure'],
+            Mercure\MercureOptions::class => static function (ContainerInterface $c): Mercure\MercureOptions {
+                $config = $c->get('config.mercure');
+                return new Mercure\MercureOptions(
+                    publicHubUrl: $config['public_hub_url'] ?? null,
+                    internalHubUrl: $config['internal_hub_url'] ?? null,
+                    jwtSecret: $config['jwt_secret'] ?? null,
+                    jwtIssuer: $config['jwt_issuer'] ?? 'Shlink',
+                );
+            },
         ],
     ],